requirements.txt
pandas~=1.4.3
python-dotenv~=0.20.0
PyMySQL~=1.0.2
SQLAlchemy~=1.4.40